requirements.txt
# Core dependencies
rapidfuzz>=3.0.0
watchdog>=3.0.0

# Database
# SQLite is built into Python

# CLI
click>=8.1.0

# Utilities
python-dateutil>=2.8.0

# Desktop UI (Phase 3)
flask>=3.0.0
flask-cors>=4.0.0
pywebview>=5.0.0

# Phase 4: System Integration
pystray>=0.19.0
pillow>=10.0.0
keyboard>=0.13.5
